backend/requirements.txt
asgiref==3.8.1
channels>=4.2.0
channels-redis>=4.0
Django>=5.1.6
sqlparse==0.5.3
typing_extensions==4.12.2
daphne>=4.0
Pillow
psycopg2-binary
whitenoise
djangorestframework